+++ Concurrent/Source/someWriter.h
@@ -40,9 +40,9 @@ public:
    unsigned short mIdent;
    unsigned short mCode;
 
-   unsigned  mPassCount;
-   unsigned  mFailCount;
-   unsigned  mCodeSum;
+   unsigned long long  mPassCount;
+   unsigned long long  mFailCount;
+   unsigned long long  mCodeSum;
 };
 
 //******************************************************************************
